% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linha_estatica.R
\name{linha_estatica}
\alias{linha_estatica}
\title{Uma função que retorna uma visualização de linha do tempo estática}
\usage{
linha_estatica(
  base_dados,
  datas,
  posicao_eixo,
  cor_bolinhas,
  primeira_informacao,
  segunda_informacao,
  titulo,
  nomex,
  nomey,
  legenda
)
}
\arguments{
\item{data}{as input to time series}
}
\value{
retorna a visualização p
}
\description{
Uma função que retorna uma visualização de linha do tempo estática
}
\examples{
linha_estatica(notten)
}
