% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linha_interativa.R
\name{linha_interativa}
\alias{linha_interativa}
\title{Uma função que arruma uma objeto plotly para torná-lo mais visual}
\usage{
linha_interativa(objetoplotly)
}
\arguments{
\item{data}{as input to time series}
}
\value{
retorna a visualização interativa
}
\description{
Uma função que arruma uma objeto plotly para torná-lo mais visual
}
\examples{
linha_interativa(notten)
}
