% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tempo.R
\name{viz_tempo}
\alias{viz_tempo}
\title{Uma função que retorna uma visualização de linha do tempo interativa diretamente através dos dados}
\usage{
viz_tempo(
  base_dados,
  datas,
  posicao_eixo,
  cor,
  primeira_informacao,
  segunda_informacao,
  titulo,
  nomex,
  nomey,
  legenda
)
}
\arguments{
\item{data}{as input to time series}
}
\value{
retorna a visualização interativa direto
}
\description{
Uma função que retorna uma visualização de linha do tempo interativa diretamente através dos dados
}
\examples{
linha_interativa(notten)
}
